add "ewl/both" as possible source, this is triggerred if the state is added to the widget or to the parent and it was not set before.
git-svn-id: http://svn.enlightenment.org/svn/e@41190 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ branches/ewl/new_state_system/src/lib/ewl_widget.h
@@ -156,9 +156,9 @@ struct Ewl_Widget
         Ecore_Hash *theme;          /**< Overriding theme settings */
         Ewl_Pair_List theme_text;   /**< Overriding text in theme */
         unsigned int flags;         /**< the widget flags */
-        unsigned int states;        /**< Direct widget states */
-        unsigned int inherited_states;    /**< States that are applied to the
-                                                theme */
+        unsigned short states;        /**< Direct widget states */
+        unsigned short inherited_states;  /**< States that are applied to 
+                                                the theme */
         unsigned char swallowed:1;
 };
